Backend/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
fastapi==0.94.0
h11==0.14.0
idna==3.4
jwt==1.3.1
passlib==1.7.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.6
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.26.0.post1
typing==3.7.4.3
typing_extensions==4.5.0
uvicorn==0.21.0
python-dotenv==1.0.0
